+++ CHANGELOG.md
@@ -21,6 +21,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Build issue with new Bootstrap releases (#152)
+- Authorization headers are now hidden on HTTP debug logs
 
 ## [2.0.0-beta1] - 2015-08-26